+++ src/algorithms/index.js
@@ -6,6 +6,7 @@ import * as cluster from './cluster';
 import * as dag from './dag';
 import * as graphical from './graphical';
 import * as isomorphism from './isomorphism';
+import * as operators from './operators';
 import * as shortestPaths from './shortestPaths';
 
 export {
@@ -15,6 +16,7 @@ export {
   dag,
   graphical,
   isomorphism,
+  operators,
   shortestPaths
 };
 
@@ -24,4 +26,5 @@ export * from './cluster';
 export * from './dag';
 export * from './graphical';
 export * from './isomorphism';
+export * from './operators';
 export * from './shortestPaths';